+++ modules/home-manager/desktop/default.nix
@@ -2,6 +2,7 @@
   imports = [
     ./gtk.nix
     ./theme
+    ./themes/monokai-pro-skeudos.nix
     ./x11
   ];
 }